// src/rest/model/deliveryLoan.ts
/**
 * Pakkasmarja REST API
 * REST API for Pakkasmarja
 *
 * OpenAPI spec version: 3.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DeliveryLoanItem } from './deliveryLoanItem';


export interface DeliveryLoan { 
    item: DeliveryLoanItem ;
    loaned: number ;
    returned: number ;
}    

export interface DeliveryLoanOpt { 
    item?: DeliveryLoanItem;
    loaned?: number;
    returned?: number;
}